+++ runner/rules/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 **Bug Fixes**
 
+* Make GrantPermissionRule use UiAutomation to grant permissions on APIs > =28 to fix issues running on automotive.
+
 **New Features**
 
 **Breaking Changes**